+++ src/main/java/act/controller/bytecode/ControllerByteCodeScanner.java
@@ -109,7 +109,6 @@ public class ControllerByteCodeScanner extends AppByteCodeScannerBase {
 
         @Override
         public void visit(int version, int access, String name, String signature, String superName, String[] interfaces) {
-            logger.trace("Scanning %s", name);
             classInfo.className(name);
             String className = name.replace('/', '.');
             if (router.possibleController(className)) {