+++ testapps/GHIssues/src/main/java/ghissues/Gh869.java
@@ -1,0 +1,12 @@
+package ghissues;
+
+import act.controller.annotation.UrlContext;
+import org.osgl.mvc.annotation.GetAction;
+
+@UrlContext("api/v1/869")
+public class Gh869 extends BaseController {
+
+    @GetAction
+    public void test() {}
+
+}